refactor: reorganize component structure
- Move homepage components to src/components/Homepage/
- Move about components to src/components/About/
- Add CI pipeline configuration with GitHub Actions
- Update imports to reflect new component locations
- Add TypeScript and ESLint checks to CI workflow
