+++ src/components/About/aboutHero1/index.tsx
