+++ src/components/About/aboutHero2/index.tsx
