+++ src/components/About/aboutHero3/index.tsx
